+++ smart-transit-app/.eslintrc.json
