+++ smart-transit-app/README.md
