+++ smart-transit-app/app/favicon.ico
