+++ smart-transit-app/app/globals.css
