+++ smart-transit-app/app/layout.tsx
