+++ smart-transit-app/next.config.mjs
