+++ smart-transit-app/package.json
