+++ smart-transit-app/tailwind.config.ts
