+++ smart-transit-app/tsconfig.json
